fixed error with deployment schedule acceleration
git-svn-id: svn://107.170.10.188:8080/Semoss/dev@5495 ef391c1c-ee15-45f7-b1d6-dde3d27a2169

Former-commit-id: 19f26dd21f620bda63eefb27fffe48ad2e1737b7

+++ src/prerna/ui/main/listener/specific/tap/DHMSMDeploymentStrategyRunBtnListener.java
@@ -279,7 +279,7 @@ public class DHMSMDeploymentStrategyRunBtnListener implements ActionListener {
 					} else {
 						distanceInQuarters = Math.abs(beginQuarter - endQuarter);
 					}
-					double numQuartersPerWave = distanceInQuarters/wavesInRegion.size();
+					double numQuartersPerWave = (double) distanceInQuarters/wavesInRegion.size();
 	
 					double currQuarter = beginQuarter;
 					int currYear = beginYear;
